# api.yaml
openapi: 3.0.0
info:
  title: simple-api
  version: 0.1.0
...
paths:
...
  /temperature:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 1
                currentTemperature:
                  type: number
                  example: 20.1
                lastUpdated:
                  type: string
                  format: local-date-time
                  example: "2024-08-20T22:55:04.425"
      responses:
        "200":
          description: Создать сенсор
          content:
            application/json:
              schema:
                type: object
                example:
                  id: 1
                  currentTemperature: 0.0
                  lastUpdated: "2024-08-20T22:55:04.425"
        "400":
          description: Не удалось создать сенсор
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error_type"
              example:
                timestamp: "2024-08-18T20:58:05.337+00:00"
                status: 400
                error: "Bad Request"
                path: "/api/temperature"
        "500":
          description: Не удалось создать сенсор
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error_type"
              example:
                timestamp: "2024-08-18T20:58:05.337+00:00"
                status: 500
                error: "Internal Server Error"
                path: "/api/temperature"
  /temperature/{id}/current-temperature:
    get:
      security:
        - bearerAuth: []
      summary: Получение текущей температуры сэнсора.
      parameters:
        - in: path
          name: id
          description: Идентификатор сенсора
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Текущая температура
          content:
            application/json:
              schema:
                type: number
        "401":
          description: Ошибка авторизации
          content:
            plain/text:
              schema:
                type: string
        "500":
          description: Ошибка получения температуры
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error_type"
              example:
                timestamp: "2024-08-18T20:58:05.337+00:00"
                status: 500
                error: "Internal Server Error"
                path: "/api/temperature/2/current-temperature"
  /temperature/{id}/set-temperature:
    post:
      summary: Установка температуры сенсора.
      parameters:
        - in: path
          name: id
          description: Идентификатор сенсора
          required: true
          schema:
            type: integer
          example: 0
        - in: query
          name: temperature
          description: Температура
          required: true
          schema:
            type: number
          example: 18.2
      responses:
        "204":
          description: Температура обновлена
        "400":
          description: Не верный формат температуры
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error_type"
              example:
                timestamp: "2024-08-18T20:58:05.337+00:00"
                status: 400
                error: "Bad Request"
                path: "/api/temperature/2/set-temperature"
        "500":
          description: Ошибка обновления температуры
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error_type"
              example:
                timestamp: "2024-08-18T20:58:05.337+00:00"
                status: 500
                error: "Internal Server Error"
                path: "/api/temperature/2/set-temperature"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purpose
  schemas:
    error_type:
      type: object
      properties:
        timestamp:
          type: string
          description: Штамп времени
          example: "2024-08-18T20:53:11.329+00:00"
        status:
          type: integer
          description: Статус
          example: 500
        error:
          type: string
          description: Текст ошибки
          example: "Internal Server Error"
        path:
          type: string
          description: Путь сервиса
          example: "/api/heating/200"
...